Add image scan to pr, release, and main-build workflows (#239)
*Description of changes:*

- Adds scanning to the ADOT Docker image created in `build-arm`,
`build-windows`,
[build-x64-musl](https://github.com/aws-observability/aws-otel-dotnet-instrumentation/actions/runs/16976441426/job/48126489009#logs)
and
[build-arm-musl](https://github.com/aws-observability/aws-otel-dotnet-instrumentation/actions/runs/16976441426/job/48126488997#logs)
steps.

- Adds a few exclusions to the `image_scan` action for dependencies and
tools used to build the local ADOT image. These dependencies are only
present in the PR and Main build workflows, so only the actual local
ADOT image that we will eventually release is scanned, and not the SDKs
and tools used to build the image itself. This parities with the
existing `daily_scan` workflow, which only scans the public ADOT ECR
image that is published.

- Rename the local images built to reflect a more accurate name for the
Docker image.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.