Change CodeQL to advanced (#257)
The default CodeQL configuration does not run for forked PRs, causing
PRs to simply get stuck. In this PR I have enabled "Advanced
configuration", and made no changes to the templated workflow that was
created. This is a 2-way door, we can always go back to default if this
proves to be the wrong approach. See
https://github.com/aws-observability/aws-otel-python-instrumentation/pull/210

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.