Update README for pending AWS SDK V4 support (#275)
*Issue #, if available:*

*Description of changes:*
This package currently has a known issue that breaks support for AWS SDK
for V4. As we are still working on a fix to support V4, update README to
inform users of this incompatibility.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: ADOT Patch workflow <adot-patch-workflow@github.com>