Update daily-scan.yml to run every 6 hours (#274)
We still see periodic failures in this workflow and sometimes 2 failures
in a row is enough to cause an alarm (b/c of when the tests run and when
the alarm is evaluated). Increase run rate to every 6 hours to reduce
chance of 2 failures causing an alarm.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.